[dev-3279] extracting Chart to its own component AwardAmountsChart
